# web/src/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "API for the Rack Gateway administration and proxy services.",
        "title": "Rack Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Returns every user configured in the gateway along with role assignments.",
                "tags": [
                    "Users"
                ],
                "summary": "List all gateway users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/db.User"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    },
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "Creates a new gateway user with the provided roles.",
                "tags": [
                    "Users"
                ],
                "summary": "Create a user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/handlers.CreateUserRequest"
                            }
                        }
                    },
                    "description": "User payload",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.UserSummary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{email}": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Returns details for a single gateway user.",
                "tags": [
                    "Users"
                ],
                "summary": "Get a user",
                "parameters": [
                    {
                        "description": "User email",
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/db.User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "SessionCookie": []
                    },
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "Updates a user's display name and/or email.",
                "tags": [
                    "Users"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "description": "Current email",
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/handlers.UpdateUserProfileRequest"
                            }
                        }
                    },
                    "description": "User profile",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.UserSummary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionCookie": []
                    },
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "Removes a gateway user and revokes all sessions they own.",
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "description": "User email",
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "components": {
...
        "securitySchemes": {
            "CSRFToken": {
                "description": "HMAC-derived CSRF token tied to the active session.",
                "type": "apiKey",
                "name": "X-CSRF-Token",
                "in": "header"
            },
            "SessionCookie": {
                "description": "HttpOnly session cookie issued after OAuth login.",
                "type": "apiKey",
                "name": "Cookie",
                "in": "header"
            }
        },
        "schemas": {
...
            "db.User": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "created_by_email": {
                        "type": "string"
                    },
                    "created_by_name": {
                        "type": "string"
                    },
                    "created_by_user_id": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "locked_at": {
                        "type": "string"
                    },
                    "locked_by_email": {
                        "type": "string"
                    },
                    "locked_by_name": {
                        "type": "string"
                    },
                    "locked_by_user_id": {
                        "type": "integer"
                    },
                    "locked_reason": {
                        "type": "string"
                    },
                    "mfa_enforced_at": {
                        "type": "string"
                    },
                    "mfa_enrolled": {
                        "type": "boolean"
                    },
                    "name": {
                        "type": "string"
                    },
                    "preferred_mfa_method": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "suspended": {
                        "type": "boolean"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
...
            "handlers.CreateUserRequest": {
                "type": "object",
                "required": [
                    "email",
                    "name",
                    "roles"
                ],
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "minItems": 1,
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "handlers.ErrorResponse": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            },
...
            "handlers.UpdateUserProfileRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}